changed name of herCinema to HerCDN
